refactor(desktop): group backend management modules
