--- desktop/src/unavailable.html
@@ -1,74 +1,0 @@
-<!doctype html>
-<html lang="zh-CN">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="default-src 'none'; style-src 'unsafe-inline'; script-src 'unsafe-inline'"
-    />
-    <title>qwen-audio-agent</title>
-    <style>
-      :root {
-        color: #f5f5f7;
-        background: #08090d;
-        font-family: Inter, "PingFang SC", "Microsoft YaHei", sans-serif;
-      }
-      * { box-sizing: border-box; }
-      body {
-        min-height: 100vh;
-        margin: 0;
-        display: grid;
-        place-items: center;
-        padding: 32px;
-        text-align: center;
-        background:
-          radial-gradient(circle at 50% 10%, #694fd43d, transparent 40%),
-          #08090d;
-      }
-      main { max-width: 320px; }
-      .mark {
-        width: 72px;
-        height: 72px;
-        margin: 0 auto 24px;
-        display: grid;
-        place-items: center;
-        border-radius: 24px;
-        font-size: 28px;
-        font-weight: 700;
-        background: linear-gradient(145deg, #a38aff, #6343e5);
-        box-shadow: 0 0 42px #7156e768;
-      }
-      h1 { margin: 0 0 10px; font-size: 24px; }
-      p { color: #918d9d; line-height: 1.7; }
-      code { color: #c5baff; }
-      button {
-        margin-top: 12px;
-        padding: 10px 18px;
-        color: white;
-        border: 1px solid #8069ef66;
-        border-radius: 10px;
-        background: #7358e8;
-        cursor: pointer;
-      }
-    </style>
-  </head>
-  <body>
-    <main>
-      <div class="mark">V</div>
-      <h1>qwen-audio-agent 尚未连接</h1>
-      <p>
-        请先启动 qwen-audio-agent 服务：<br />
-        <code>qwenaudio gateway</code>
-      </p>
-      <button type="button" id="retry">重新连接</button>
-    </main>
-    <script>
-      const target = new URLSearchParams(location.search).get('target')
-        || 'http://127.0.0.1:3101'
-      document.querySelector('#retry').addEventListener('click', () => {
-        location.replace(target)
-      })
-    </script>
-  </body>
-</html>
