+++ desktop/test/backend-configuration.test.mjs
@@ -4,7 +4,7 @@ import test from 'node:test'
 import {
   configurationLaunch,
   openBackendConfiguration,
-} from '../src/backend-configuration.mjs'
+} from '../src/backend/configuration.mjs'
 import { backendNames } from '../../shared/backend-catalog.mjs'
 
 test('builds native terminal launches for macOS and Windows', () => {
@@ -15,6 +15,36 @@ test('builds native terminal launches for macOS and Windows', () => {
   const windows = configurationLaunch('codex', { env: {}, platform: 'win32' })
   assert.equal(windows.command, 'cmd.exe')
   assert.ok(windows.args.includes('codex login'))
+})
+
+test('builds fixed backend-owned configuration commands', () => {
+  const qoder = configurationLaunch('qoder', { env: {}, platform: 'darwin' })
+  assert.match(qoder.args.join(' '), /qodercli login/)
+
+  const codebuddy = configurationLaunch('codebuddy', {
+    env: {},
+    platform: 'darwin',
+  })
+  assert.match(codebuddy.args.join(' '), /codebuddy/)
+})
+
+test('maps trusted configuration actions on Linux and Windows', () => {
+  const linux = configurationLaunch('deepseek', {
+    env: {},
+    platform: 'linux',
+  })
+  assert.equal(linux.command, 'x-terminal-emulator')
+  assert.deepEqual(linux.args.slice(0, 3), ['-e', 'sh', '-lc'])
+  assert.match(linux.args[3], /^dsh web;/)
+
+  const windows = configurationLaunch('codex', {
+    env: {},
+    platform: 'win32',
+  })
+  assert.equal(windows.command, 'cmd.exe')
+  assert.deepEqual(windows.args, [
+    '/d', '/c', 'start', '', 'cmd.exe', '/k', 'codex login',
+  ])
 })
 
 test('maps every backend-owned configuration action on every desktop OS', () => {
@@ -51,6 +81,27 @@ test('falls back across common Linux terminal emulators', async () => {
   })
   assert.equal(result.ok, true)
   assert.deepEqual(calls, ['x-terminal-emulator', 'gnome-terminal'])
+})
+
+test('launches configuration detached from the desktop process', async () => {
+  const calls = []
+  let unref = false
+  const result = await openBackendConfiguration('codex', {
+    env: {},
+    platform: 'darwin',
+    spawnImpl: (...args) => {
+      calls.push(args)
+      const child = new EventEmitter()
+      child.unref = () => { unref = true }
+      queueMicrotask(() => child.emit('spawn'))
+      return child
+    },
+  })
+  assert.equal(result.ok, true)
+  assert.equal(result.action.kind, 'terminal')
+  assert.match(calls[0][1].join(' '), /codex login/)
+  assert.equal(calls[0][2].detached, true)
+  assert.equal(unref, true)
 })
 
 test('reports a clear Linux error when no terminal exists', async () => {
